+++ framework/caching/FileCache.php
@@ -107,11 +107,19 @@ class FileCache extends Cache
     protected function getValue($key)
     {
         $cacheFile = $this->getCacheFile($key);
+
         if (@filemtime($cacheFile) > time()) {
-            return @file_get_contents($cacheFile);
-        } else {
-            return false;
+            $fp = @fopen($cacheFile, "r");
+            if ($fp !== false) {
+                @flock($fp, LOCK_SH);
+                $cacheValue = @file_get_contents($cacheFile);
+                @flock($fp, LOCK_UN);
+                @fclose($fp);
+                return $cacheValue;
+            }
         }
+
+        return false;
     }
 
     /**